+++ include/subsystemHeaders/drive.hpp
@@ -1,3 +1,5 @@
 #include "api.h"
 
+extern void getMotorOutput();
+extern void DirectionToggle();
 extern void setDriveMotors();