+++ src/main.cpp
@@ -1,4 +1,5 @@
 #include "main.h"
+#include "pros/motors.h"
 
 /**
  * A callback function for LLEMU's center button.
@@ -24,9 +25,9 @@ void on_center_button() {
  */
 void initialize() {
 	pros::lcd::initialize();
-	pros::lcd::set_text(1, "Hello PROS User!");
+	// pros::lcd::set_text(1, "Hello PROS User!");
 
-	pros::lcd::register_btn1_cb(on_center_button);
+	// pros::lcd::register_btn1_cb(on_center_button);
 }
 
 /**
@@ -35,3 +36,25 @@ void initialize() {
  * the robot is enabled, this task will exit.
  */
 void disabled() {}
+
+/**
+ * Runs after initialize(), and before autonomous when connected to the Field
+ * Management System or the VEX Competition Switch. This is intended for
+ * competition-specific initialization routines, such as an autonomous selector
+ * on the LCD.
+ *
+ * This task will exit when the robot is enabled and autonomous or opcontrol
+ * starts.
+ */
+void competition_initialize() {
+	CIDisplay();
+	IntakeMotor.set_brake_mode(pros::E_MOTOR_BRAKE_BRAKE);
+	while (true) {
+		if(Controller.get_digital_new_press(DIGITAL_DOWN)) DownPressed();
+		if(Controller.get_digital_new_press(DIGITAL_UP)) UpPressed();
+		if(Controller.get_digital_new_press(DIGITAL_A)) APressed();
+		if(Controller.get_digital_new_press(DIGITAL_B)) BPressed();
+
+		pros::delay(20);
+	}
+}