+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v1.0.0 (12/17/22) - Capital Beltway Challenge
+- Added input curving to the motor
+- Merged competition_initialize.cpp back into main.cpp
+- Added delays between controller printing to make controller display functions work
+- Fixed bug with AutonSelect display occuring when checking motor temperatures
+- Update to PROS CLI 3.4.1
+- Added keybind checking and notes to the pre-auton menu
+- Created auton drive & turn functions
+- Driver can now switch which side is the "head" of the bot
+- Created Driver Controller Display: displays the current drive direction and flywheel spin status
+- Added intake, indexer, and flywheel motors to the motor check page
+----------------------------------------------------------------------------------------------------------------------------------
 v0.1.1 (10/21/22)
 - Created CHANGELOG.md
 - Created drive.hpp, drive.cpp, competition_initialize.cpp, autonomous.cpp, & opcontrol.cpp